删除合并的学期服务
