+++ evaluation-service/calculation-service/pom.xml
@@ -71,7 +71,7 @@
     </dependencies>
 
     <build>
-        <finalName>evaluation-analysis-service</finalName>
+        <finalName>evaluation-calculation-service</finalName>
         <plugins>
             <plugin>
                 <groupId>org.springframework.boot</groupId>
