+++ evaluation-service/calculationrules-service/pom.xml
@@ -106,7 +106,7 @@
     </dependencies>
 
     <build>
-        <finalName>evaluation-class-service</finalName>
+        <finalName>evaluation-calculationrule-service</finalName>
         <plugins>
             <plugin>
                 <groupId>org.springframework.boot</groupId>
