+++ evaluation-service/student-evaluation/pom.xml
@@ -106,7 +106,7 @@
     </dependencies>
 
     <build>
-        <finalName>evaluation-studentevaluation-service</finalName>
+        <finalName>student-evaluation-service</finalName>
         <plugins>
             <plugin>
                 <groupId>org.springframework.boot</groupId>
