+++ evaluation-service/teacher-evaluation/pom.xml
@@ -85,7 +85,7 @@
     </dependencies>
 
     <build>
-        <finalName>evaluation-teacher-evaluation-service</finalName>
+        <finalName>teacher-evaluation-service</finalName>
         <plugins>
             <plugin>
                 <groupId>org.springframework.boot</groupId>
